Fix duplication of OBF Update talk slot.
This fix would be presuming that the OBF President's Update is indeed now taking place on the first day, and not the morning of the second day. If it has been moved back to the morning of the second day, the entry on the first day needs to be removed instead (and then reject this PR).